fix(x64): widen switchbot packet vnum/count; fix qc token-set typo
- packet.h TSwitchbotUpdateItem: vnum/count BYTE -> DWORD. Verified
  against the full system: the server fills them from DWORD GetVnum()/
  GetCount(); the client reads count via SetItemCount (item stacks can
  exceed 255, so the BYTE truncation was a real bug). Shared struct, so
  client and server stay in sync on rebuild. (vnum is not read by the
  client today, but is now correctly typed.)
- qc.cc: 'prev.t.token == TK_DO' -> '= TK_DO' (copy-paste typo; the
  identical sibling block at line 730 uses '='; prev.t.token is read
  downstream so the assignment is meaningful).

Server x64 now builds with 0 compile errors and 0 warnings (/W3, no
suppressions). Remaining 84 errors are link-only (x64 dependency libs).

+++ Lead-Shared-Source/packet.h
@@ -2321,8 +2321,8 @@ struct TPacketGCSwitchbot
 struct TSwitchbotUpdateItem
 {
 	BYTE	slot;
-	BYTE	vnum;
-	BYTE	count;
+	DWORD	vnum;
+	DWORD	count;
 	long	alSockets[ITEM_SOCKET_MAX_NUM];
 	TPlayerItemAttribute aAttr[ITEM_ATTRIBUTE_MAX_NUM];
 };
